split header into components
