+++ src/components/nav/SearchOverlay.jsx
@@ -1,6 +1,6 @@
 import React from "react";
-import searchIconThin from "/public/assets/icons/magnifying-glass-thin.svg";
-import backIcon from "/public/assets/icons/back-white.svg";
+import searchIconThin from "/src/assets/icons/magnifying-glass-thin.svg";
+import backIcon from "/src/assets/icons/back-white.svg";
 
 function SearchOverlay({ onClose, handleSearchSubmit, searchFormData, handleSearchChange }) {
   return (
@@ -8,7 +8,7 @@ function SearchOverlay({ onClose, handleSearchSubmit, searchFormData, handleSear
       <div className="search-overlay-content">
         <form onSubmit={handleSearchSubmit} className="input-container">
           <img className="back-icon" src={backIcon} onClick={onClose} />
-          <input type="text" className="overlay-search-input secondary-green-colour text-light" placeholder="Search..." value={searchFormData} onChange={handleSearchChange} />
+          <input type="text" className="overlay-search-input secondary-green-colour text-light" placeholder="Search..." value={searchFormData} onChange={handleSearchChange} aria-label="Search field" />
           <button className="search-icon">
             <img className="search-icon" src={searchIconThin} />
           </button>
