+++ src/components/nav/SearchformLarge.jsx
@@ -1,13 +1,13 @@
 import React from "react";
-import searchIconThin from "/public/assets/icons/magnifying-glass-thin.svg";
+import searchIconThin from "/src/assets/icons/magnifying-glass-thin.svg";
 
 function SearchFieldLargeScreen({ handleSearchSubmit, searchFormData, handleSearchChange }) {
   return (
     <div className="container d-none d-lg-block position-relative">
       <form onSubmit={handleSearchSubmit}>
         <input type="text" className="search-form-input w-100 secondary-green-colour" placeholder="Search for a product e.g. milk og crowbars" value={searchFormData} onChange={handleSearchChange} />
-        <button className="search-icon">
-          <img src={searchIconThin} />
+        <button className="search-icon" aria-label="Search field">
+          <img src={searchIconThin} alt="Logo" />
         </button>
       </form>
     </div>
